Hilton Room Modifications
A fun project where i took photos of a Hilton Hotel room I was staying in and used AI to generate images of the room using modern settings in day and night time. I have applied color combinations, furniture themes, artefacts and tried generating the images during day and night both. 

I reached out to the Hilton Hotel manager to share this idea with her, unfortunately it was her last day at Hilton so did not get any traction. Will keep looking for contacts in Hilton to pitch these images to the right folks so that they can take it forward (04/30).