IoT Marketing Strategy for a Small Consulting Firm
An IoT strategy that I developed for a small consulting company that had a meeting planned with the CEO of the company. Used the best of Manus, Gemini, ChatGPT, Notebook LLM and my own collaborative skills to get all the docs out. The main one is the IoT Strategy* pdf doc that was used for the discussion.