Simple newsletter project
Here is a fun and easy way to be on top of the latest news, articles and happenings in the world. As I'm deep diving into the Agentic Automation frameworks  I made the attached news letter for this space. If you would like to get subscribed to this newsletter (will be sent every 3 days) please dm or update your email address here: http://bit.ly/4iKFkr1.



TechDetais: Here I'm using gpt-4o-search (paid version) to do live search that is acting as an assistant to Gemini 2.5 LLM model. Now this is simply superb as we know LLMs on their own cannot do live search and are trained at best to the last date data that they were trained on. This removes that blocker and makes searching the web for live information easy and accurate. Another key point to note here is that the heavy lifting that is generally done by the LLM now is being done by the Tool (gpt-4o-search in this case) so that frees up the LLM (Gemni) in this case to curate the content and present it to the subscriber. Let me know if you have any questions, had lot of fun doing this. 



Next Steps: This could be a newsletter for any industry, any niche area. This is just one use-case it can be easily extended to any industry, any niche area, it saves time and is customized for the user.